backend/requirements.txt
Flask
flask-cors
pdfplumber
python-docx
python-dotenv
